+++ Assets/ProceduralStages/CHANGELOG.md
@@ -1,3 +1,9 @@
+## 1.17.3
+1. Fixed lunar golems not spawning
+2. Fixed props desync
+3. Fixed moon desync
+4. Use moon music on moon
+
 ## 1.17.2
 1. Fixed lag spikes during the Mithrix fight by simplifying the air node graph. This should also improve the performance of pathfinding for flying enemies.